Convert NoteCard to a class based component
- With the intended use of a model, NoteCard will need to handle the
form's submission.
- From this intended use, NoteCard will no longer be
purely presentational, hence the decision to make it a class based
component
- While porting over to class based component, a copy-paste error was
also identified. The card on which the note's details appear supports a
maximum string length of 42. The intention was to truncate strings
that exceed this length using slice(0, 38) and follow this substring
with a ` ...`. Due to a copy and paste error, this operation was not
being performed on the string's description.